Converted sp->planar, pmid and iscale from glong to greal.

+++ xyplot.c
@@ -92,8 +92,8 @@ xy_reproject (splotd *sp, greal **world_data, datad *d, ggobid *gg)
   for (i=0; i<d->nrows; i++) {
     m = d->rows_in_plot[i];
 
-    sp->planar[m].x = (glong) world_data[m][jx];  /*-- regular or missings --*/
-    sp->planar[m].y = (glong) world_data[m][jy];
+    sp->planar[m].x = world_data[m][jx];  /*-- regular or missings --*/
+    sp->planar[m].y = world_data[m][jy];
   }
 }
